funtech/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.1.7
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
drf-spectacular==0.27.0
gunicorn==21.2.0
idna==3.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
